Revert "Revert D21089648: Put TORCH_LIBRARY in torch/library.h; add custom class API"
This reverts commit 2ccdc39dce91a1821ede9bfdea26b30d66e1554f.

Original PR: https://github.com/pytorch/pytorch/pull/36742
