+++ aten/src/ATen/native/quantized/cpu/qbatch_norm.cpp
@@ -1,7 +1,7 @@
 #include <ATen/ATen.h>
 #include <ATen/NativeFunctions.h>
 #include <ATen/Parallel.h>
-#include <ATen/core/op_registration/op_registration.h>
+#include <torch/library.h>
 #include <ATen/native/quantized/cpu/quantized_ops.h>
 
 #include <algorithm>
