+++ aten/src/ATen/native/xnnpack/RegisterOpContextClass.cpp
@@ -1,6 +1,6 @@
 #ifdef USE_XNNPACK
 
-#include <ATen/core/op_registration/op_registration.h>
+#include <torch/library.h>
 #include <ATen/native/xnnpack/Convolution.h>
 #include <ATen/native/xnnpack/Linear.h>
 #include <ATen/native/xnnpack/OpContext.h>
