+++ aten/src/ATen/templates/BackendSelectRegister.cpp
@@ -6,7 +6,7 @@
 
 #include <ATen/ATen.h>
 #include <ATen/Dispatch.h>
-#include <ATen/core/op_registration/op_registration.h>
+#include <torch/library.h>
 #include <c10/core/TensorOptions.h>
 
 namespace at {
