+++ aten/src/ATen/templates/PerOpRegistration.cpp
@@ -1,7 +1,7 @@
 // ${generated_comment}
 
 #include <ATen/Config.h>
-#include <ATen/core/op_registration/op_registration.h>
+#include <torch/library.h>
 #include <ATen/TypeDefault.h>
 $extra_headers
