+++ aten/src/ATen/templates/SchemaRegister.cpp
@@ -1,7 +1,7 @@
 // ${generated_comment}
 
 #include <c10/core/TensorOptions.h>
-#include <ATen/core/op_registration/op_registration.h>
+#include <torch/library.h>
 
 using namespace at;
