+++ aten/src/ATen/templates/TypeDefault.cpp
@@ -13,7 +13,7 @@
 #include <c10/core/TensorOptions.h>
 #include <ATen/DeviceGuard.h>
 #include <ATen/SparseTensorUtils.h>
-#include <ATen/core/op_registration/op_registration.h>
+#include <torch/library.h>
 
 namespace {
 static const char* named_tensors_unsupported_error =
