+++ aten/src/ATen/test/extension_backend_test.cpp
@@ -2,7 +2,7 @@
 
 #include <ATen/ATen.h>
 #include <ATen/NativeFunctions.h>
-#include <ATen/core/op_registration/op_registration.h>
+#include <torch/library.h>
 
 #include <torch/csrc/jit/runtime/operator.h>
