+++ caffe2/CMakeLists.txt
@@ -699,7 +699,12 @@ endif()
   install(DIRECTORY "${TORCH_SRC_DIR}/csrc"
     DESTINATION ${TORCH_INSTALL_INCLUDE_DIR}/torch
     FILES_MATCHING PATTERN "*.h")
-  install(FILES "${TORCH_SRC_DIR}/script.h" "${TORCH_SRC_DIR}/extension.h" "${TORCH_SRC_DIR}/custom_class.h" "${TORCH_SRC_DIR}/custom_class_detail.h"
+  install(FILES
+    "${TORCH_SRC_DIR}/script.h"
+    "${TORCH_SRC_DIR}/extension.h"
+    "${TORCH_SRC_DIR}/custom_class.h"
+    "${TORCH_SRC_DIR}/library.h"
+    "${TORCH_SRC_DIR}/custom_class_detail.h"
     DESTINATION ${TORCH_INSTALL_INCLUDE_DIR}/torch)
