+++ torch/csrc/jit/mobile/register_mobile_autograd.cpp
@@ -1,8 +1,8 @@
 #include <ATen/ATen.h>
 #include <ATen/TypeDefault.h>
-#include <ATen/core/op_registration/op_registration.h>
 #include <ATen/core/stack.h>
 #include <torch/csrc/autograd/function.h>
+#include <torch/library.h>
 
 using Stack = std::vector<c10::IValue>;
 using at::Scalar;
@@ -104,12 +104,12 @@ void log_softmax_kernel(const c10::OperatorHandle& op, Stack* stack) {
 TORCH_LIBRARY_IMPL(_aten, Autograd, m) {
   m.impl("add.Scalar", torch::autograd::VariableType::add_Scalar);
   m.impl("mul.Tensor", torch::autograd::VariableType::mul_Tensor);
-  m.impl("conv2d", CppFunction::makeFromBoxedFunction<conv2d_kernel>());
+  m.impl("conv2d", torch::CppFunction::makeFromBoxedFunction<conv2d_kernel>());
   m.impl("dropout", VariableType::dropout);
   m.impl("feature_dropout", VariableType::feature_dropout);
   m.impl(
       "log_softmax.int",
-      CppFunction::makeFromBoxedFunction<log_softmax_kernel>());
+      torch::CppFunction::makeFromBoxedFunction<log_softmax_kernel>());
   m.impl(
       "max_pool2d",
       [](const Tensor& self,
@@ -127,7 +127,7 @@ TORCH_LIBRARY_IMPL(_aten, Autograd, m) {
             ceil_mode);
       });
   m.impl("relu", VariableType::relu);
-  m.impl("view", CppFunction::makeFromBoxedFunction<view_kernel>());
+  m.impl("view", torch::CppFunction::makeFromBoxedFunction<view_kernel>());
   m.impl("t", VariableType::t);
   m.impl("addmm", VariableType::addmm);
 }
